Initial commit: test for intersection via polynomials
This is a very basic demonstration of the idea for Private Set Intersection. Take one set, turn it into the coefficients of a polynomial that roots are the elements of the set. Evaluate the polynomial on the elements of the adversial set, and add the original number to it. If the element is present in the adversial set, the polynomial will evaluate to zero, so adding the original number to it will produce a recognisable number for the requester side. This is just the intersection but without encryption the it's not really private and it is prone to produce false positives.

This is just to start the project itself.