refactor: introducing points for change the polynomial type
The PSI algorithm is using Additive ElGamal encryption system, which is basically just using the g^m (g = generator, m = message) instead of g. This way, the message will not be completely decipherable, but identity is enough.

In this form, we only interested in the Polynomial to accept a Field type, which in this case will be `int`, but later this will be replaced with the cipher type of the Additive ElGamal crypto-system.