refactor: remove the Order and Generator as runtime arguments
Generally the order and the generator for the field can be agreed ahead of the transaction. It is only the key pair that needs to be provided during runtime.

Using a templated struct would allow to parameterize the ElGamal 'namespace'. However this comes at a price that I need to move operators within their respective scope, instead of letting them be global.